+++ src/apyfloatarray.cc
@@ -23,6 +23,7 @@ namespace nb = nanobind;
 #include <stdexcept>
 #include <stdlib.h>
 #include <string>
+#include <string_view>
 #include <variant>
 
 void APyFloatArray::create_in_place(
@@ -50,24 +51,31 @@ APyFloatArray::APyFloatArray(
     std::uint8_t man_bits,
     std::optional<exp_t> bias
 )
-    : APyArray(python_sequence_extract_shape(sign_seq))
+    : APyArray(python_sequence_extract_shape(sign_seq, "APyFloatArray.__init__"))
     , exp_bits(exp_bits)
     , man_bits(man_bits)
 {
-    this->bias = bias.value_or(APyFloat::ieee_bias(exp_bits));
+    constexpr std::string_view caller_name = "APyFloatArray.__init__";
 
-    const auto& signs_shape = _shape;
-    const auto exps_shape = python_sequence_extract_shape(exp_seq);
-    const auto mans_shape = python_sequence_extract_shape(man_seq);
-
+    const std::vector<std::size_t>& signs_shape = _shape;
+    const auto exps_shape = python_sequence_extract_shape(exp_seq, caller_name);
+    const auto mans_shape = python_sequence_extract_shape(man_seq, caller_name);
     if (!((signs_shape == exps_shape) && (signs_shape == mans_shape))) {
-        throw std::domain_error("Shape mismatch during construction");
+        throw std::domain_error(
+            fmt::format(
+                "APyFloatArray.__init__: shape mismatch, sign: {}, exp: {}, man: {}",
+                tuple_string_from_vec(signs_shape),
+                tuple_string_from_vec(exps_shape),
+                tuple_string_from_vec(mans_shape)
+            )
+        );
     }
 
-    auto signs = python_sequence_walk<nb::int_, nb::bool_>(sign_seq);
-    auto exps = python_sequence_walk<nb::int_>(exp_seq);
-    auto mans = python_sequence_walk<nb::int_>(man_seq);
+    auto signs = python_sequence_walk<nb::int_, nb::bool_>(sign_seq, caller_name);
+    auto exps = python_sequence_walk<nb::int_>(exp_seq, caller_name);
+    auto mans = python_sequence_walk<nb::int_>(man_seq, caller_name);
 
+    this->bias = bias.value_or(APyFloat::ieee_bias(exp_bits));
     for (std::size_t i = 0; i < signs.size(); ++i) {
         bool sign;
         if (nb::isinstance<nb::bool_>(signs[i])) {
@@ -849,31 +857,31 @@ APyFloatArray::nanmin(std::optional<std::variant<nb::tuple, nb::int_>> py_axis) 
 
 std::string APyFloatArray::repr() const
 {
-    std::stringstream ss {};
-    ss << "APyFloatArray(";
-    if (_shape[0]) {
-        std::stringstream sign_str {}, exp_str {}, man_str {};
-        sign_str << "[";
-        exp_str << "[";
-        man_str << "[";
-        for (std::size_t i = 0; i < _data.size(); ++i) {
-            const APyFloatData d = _data[i];
-            const bool is_last = i == (_data.size() - 1);
-            sign_str << (d.sign ? "1" : "0") << (is_last ? "" : ", ");
-            exp_str << d.exp << (is_last ? "" : ", ");
-            man_str << d.man << (is_last ? "" : ", ");
-        }
-        ss << sign_str.str() << "], " << exp_str.str() << "], " << man_str.str()
-           << "], ";
+    const auto sign_formatter = [](auto cbegin_it, auto _) -> std::string {
+        return fmt::format("{}", cbegin_it->sign);
+    };
+    const auto exp_formatter = [](auto cbegin_it, auto _) -> std::string {
+        return fmt::format("{}", cbegin_it->exp);
+    };
+    const auto man_formatter = [](auto cbegin_it, auto _) -> std::string {
+        return fmt::format("{}", cbegin_it->man);
+    };
+
+    if (bias == ieee_bias(exp_bits)) {
+        return array_repr(
+            exp_formatter,
+            { fmt::format("exp_bits={}", exp_bits),
+              fmt::format("man_bits={}", man_bits) }
+        );
     } else {
-        ss << "[], [], [], ";
+        return array_repr(
+            exp_formatter,
+            { /* kw_args = */
+              fmt::format("exp_bits={}", exp_bits),
+              fmt::format("man_bits={}", man_bits),
+              fmt::format("bias={}", bias) }
+        );
     }
-    ss << "shape=";
-    ss << tuple_string_from_vec(_shape);
-    ss << ", " << "exp_bits=" << static_cast<unsigned>(exp_bits) << ", "
-       << "man_bits=" << static_cast<unsigned>(man_bits) << ", " << "bias=" << bias
-       << ")";
-    return ss.str();
 }
 
 std::variant<
@@ -975,11 +983,15 @@ APyFloatArray APyFloatArray::from_numbers(
     }
 
     APyFloatArray result(
-        python_sequence_extract_shape(number_seq), exp_bits, man_bits, bias
+        python_sequence_extract_shape(number_seq, "APyFloatArray.from_float"),
+        exp_bits,
+        man_bits,
+        bias
     );
 
-    const auto py_obj
-        = python_sequence_walk<nb::float_, nb::int_, APyFixed, APyFloat>(number_seq);
+    const auto py_obj = python_sequence_walk<nb::float_, nb::int_, APyFixed, APyFloat>(
+        number_seq, "APyFloatArray.from_float"
+    );
 
     for (std::size_t i = 0; i < result._data.size(); i++) {
         if (nb::isinstance<nb::float_>(py_obj[i])) {
@@ -1117,10 +1129,15 @@ APyFloatArray APyFloatArray::from_bits(
     }
 
     APyFloatArray result(
-        python_sequence_extract_shape(python_bit_patterns), exp_bits, man_bits, bias
+        python_sequence_extract_shape(python_bit_patterns, "APyFloatArray.from_bits"),
+        exp_bits,
+        man_bits,
+        bias
     );
 
-    const auto py_obj = python_sequence_walk<nb::float_, nb::int_>(python_bit_patterns);
+    const auto py_obj = python_sequence_walk<nb::float_, nb::int_>(
+        python_bit_patterns, "APyFloatArray.from_bits"
+    );
 
     APyFloat f(exp_bits, man_bits, result.bias);
     for (std::size_t i = 0; i < result._data.size(); i++) {